refactor: split DiscordObject and payload helpers into leaf modules
