+++ src/cordless/context.py
@@ -2,7 +2,15 @@ import base64
 import json
 
 from ._multipart import build_multipart_body
-from .errors import MessageTooLongError
+from ._payload import (
+    _FLAG_EPHEMERAL,
+    _FLAG_UI_KIT,
+    _attach_files,
+    _contains_uikit,
+    _validate_content_length,
+    _validate_uikit,
+    _with_guild_id,
+)
 from .models import Attachment, Channel, Guild, Member, Message, Role, User, _wrap
 
 _CHANNEL_MESSAGE_WITH_SOURCE = 4
@@ -12,75 +20,6 @@ _DEFERRED_UPDATE_MESSAGE = 6
 _AUTOCOMPLETE_RESULT = 8
 _MODAL = 9
 
-_FLAG_EPHEMERAL = 64
-_FLAG_UI_KIT = 32768
-
-_MAX_CONTENT_LENGTH = 2000
-_MAX_UIKIT_COMPONENTS = 40
-_MAX_UIKIT_TEXT_LENGTH = 4000
-
-
-# Components v2 types: Section, TextDisplay, Thumbnail, MediaGallery, File, Separator, Container
-_UI_KIT_TYPES = {9, 10, 11, 12, 13, 14, 17}
-
-
-def _contains_uikit(components):
-    if not components:
-        return False
-    for c in components:
-        if getattr(c, "is_ui_kit", False):
-            return True
-        if isinstance(c, dict):
-            if c.get("type") in _UI_KIT_TYPES:
-                return True
-            if _contains_uikit(c.get("components")):
-                return True
-        # recurse into ActionRow children
-        elif hasattr(c, "components"):
-            if _contains_uikit(c.components):
-                return True
-    return False
-
-
-def _count_components(components):
-    """Discord counts every component in the tree toward the 40-component
-    cap, including ones nested inside a Container/Section/ActionRow and a
-    Section's accessory."""
-    if not components:
-        return 0
-    total = 0
-    for c in components:
-        total += 1
-        if isinstance(c, dict):
-            total += _count_components(c.get("components"))
-            if c.get("accessory") is not None:
-                total += 1
-        else:
-            if hasattr(c, "components"):
-                total += _count_components(c.components)
-            if getattr(c, "accessory", None) is not None:
-                total += 1
-    return total
-
-
-def _uikit_text_length(components):
-    """Sum of every TextDisplay's content, which Discord caps at 4000
-    characters total across the whole message."""
-    if not components:
-        return 0
-    total = 0
-    for c in components:
-        if isinstance(c, dict):
-            if c.get("type") == 10:
-                total += len(c.get("content") or "")
-            total += _uikit_text_length(c.get("components"))
-        else:
-            if hasattr(c, "content") and not hasattr(c, "components"):
-                total += len(c.content or "")
-            if hasattr(c, "components"):
-                total += _uikit_text_length(c.components)
-    return total
-
 
 def _leaf_options(data):
     """Descend through subcommand/group wrappers to the actual value options."""
@@ -88,45 +27,6 @@ def _leaf_options(data):
     while options and options[0].get("type") in (1, 2):
         options = options[0].get("options", [])
     return options
-
-
-def _validate_content_length(content):
-    """Discord rejects the interaction response/followup outright when
-    content is too long, but that rejection happens on Discord's end after
-    we've already returned 200 to API Gateway, so we check upfront instead
-    of failing invisibly."""
-    if content is not None and len(content) > _MAX_CONTENT_LENGTH:
-        raise MessageTooLongError(
-            f"Message content is {len(content)} characters, which exceeds Discord's {_MAX_CONTENT_LENGTH}-character limit"
-        )
-
-
-def _validate_uikit(content, embeds, components):
-    """Discord rejects a Components v2 message that also sets content or
-    embeds, or that exceeds its component-count/text-length caps, but that
-    rejection happens on Discord's end after we've already sent the request
-    (an HTTP round trip for the REST layer, or after we've already returned
-    200 to API Gateway for an interaction response), so check upfront
-    instead of failing invisibly or late. Shared between the interaction
-    path (_build_message_data below) and the REST layer's create_message/
-    edit_channel_message, so both reject the same way instead of only one
-    of them catching it locally and the other relying on Discord's own
-    validation. content/embeds must already be normalized to None by the
-    caller for "not being set in this call" (e.g. _rest/messages.py's
-    UNSET sentinel), since None here specifically means "no conflict"."""
-    if content is not None or embeds is not None:
-        raise ValueError("Components v2 messages can't also set content or embeds, use TextDisplay/Container instead")
-    count = _count_components(components)
-    if count > _MAX_UIKIT_COMPONENTS:
-        raise ValueError(
-            f"Message has {count} components, which exceeds Discord's {_MAX_UIKIT_COMPONENTS}-component limit"
-        )
-    text_length = _uikit_text_length(components)
-    if text_length > _MAX_UIKIT_TEXT_LENGTH:
-        raise MessageTooLongError(
-            f"Components v2 text totals {text_length} characters, which exceeds "
-            f"Discord's {_MAX_UIKIT_TEXT_LENGTH}-character limit"
-        )
 
 
 def _build_message_data(msg, content, embeds, components, ephemeral=False, allowed_mentions=None):
@@ -155,26 +55,6 @@ def _build_message_data(msg, content, embeds, components, ephemeral=False, allow
     if flags:
         data["flags"] = flags
     return data
-
-
-def _attach_files(data, files):
-    """Add the attachments metadata array Discord expects alongside a multipart body.
-
-    Appended after whatever's already in data["attachments"] (e.g. an edit's
-    retained-attachment list), rather than replacing it: the new entries'
-    "id" is the file's index, matching the "files[n]" part build_multipart_body
-    gives it, while retained attachments keep their own real snowflake id."""
-    existing = data.get("attachments") or []
-    data["attachments"] = existing + [{"id": i, "filename": name} for i, (name, _) in enumerate(files)]
-
-
-def _with_guild_id(data, guild_id):
-    """Discord's member and role payloads never carry their own guild_id -
-    it's implied by the endpoint you fetched them from. Stitch it in so
-    member.add_role()/role.edit() and friends know which guild to act on."""
-    if data is None or guild_id is None:
-        return data
-    return {**data, "guild_id": guild_id}
 
 
 def _wrap_members(members, users, guild_id):
